+++ test/unit/test_authorized_keys.py
@@ -3,10 +3,11 @@ Unit tests for _authorized_keys
 """
 
 import os
+import sys
 from os import path
 from unittest import TestCase, main
 
-from _authorized_keys import _FILE_NAME, AuthorizedKeys
+from authk._authorized_keys import _FILE_NAME, AuthorizedKeys
 
 
 class TestAuthorizedKey(TestCase):
@@ -15,8 +16,15 @@ class TestAuthorizedKey(TestCase):
     """
 
     def setUp(self):
-        with open(_FILE_NAME, "w+", encoding="utf-8"):
-            print(_FILE_NAME.split("/")[-1])
+        if not os.path.isdir(_FILE_NAME.replace("authorized_keys", "")):
+            os.mkdir(_FILE_NAME.replace("authorized_keys", ""))
+        os.close(os.open(_FILE_NAME, os.O_RDWR | os.O_CREAT))
+        if os.path.isfile(_FILE_NAME):
+            with open(_FILE_NAME, "w+", encoding="utf-8"):
+                print(f'{_FILE_NAME.split("/")[-1]} created')
+        else:
+            print("Something wrong happened")
+            sys.exit(-1)
 
     def tearDown(self):
         if path.exists(_FILE_NAME):
